rename setup directory to install
