+++ install/gcp/installer/util/utils.py
