+++ install/scripts/docker_setup_forseti.sh
@@ -36,9 +36,9 @@ fi
 # This assumes the script is run from the top of the source-tree.
 if [ -x "$(command -v docker)" ]; then
     echo "Building our Docker base image... "
-    docker -l error build -t forseti/base -f setup/docker/base .
+    docker -l error build -t forseti/base -f install/docker/base .
     echo "Building our Forseti image from the Docker base image... "
-    docker -l error build -t forseti/build -f setup/docker/forseti --no-cache .
+    docker -l error build -t forseti/build -f install/docker/forseti --no-cache .
 else
     echo "ERROR: Docker must be installed and it isn't, exiting." && exit 1
 fi
